serial_reader/requirements.txt
pyserial==3.3
websockets==3.3
